Trivial fixes to unit tests for gcc 4.9.2 warnings
This commit contains a collection of trivial fixes to our unit tests to
eliminate the compiler warnings on gcc 4.9.2

+++ src/PingServiceTest.cc
@@ -64,7 +64,7 @@ class MockTableConfigFetcher : public ObjectFinder::TableConfigFetcher {
                                     ObjectFinder::Indexlet>* tableIndexMap) {
 
         tableMap->clear();
-        Tablet rawEntry({tableId, 0, ~0, ServerId(),
+        Tablet rawEntry({tableId, 0, uint64_t(~0), ServerId(),
                     Tablet::NORMAL, LogPosition()});
         TabletWithLocator entry(rawEntry, locator);
